Cleanup file structure
- lib -> src
- Move index.jsx into src
- Group files into folders by domain (not by technology like components)
