+++ package.json
@@ -25,8 +25,8 @@
     "build": "nrfconnect-scripts build-prod",
     "nordic-publish": "nrfconnect-scripts nordic-publish",
     "lint-init": "nrfconnect-scripts lint-init",
-    "lint": "nrfconnect-scripts lint lib index.jsx",
-    "lintfix": "nrfconnect-scripts lint --fix lib index.jsx",
+    "lint": "nrfconnect-scripts lint src",
+    "lintfix": "nrfconnect-scripts lint --fix src",
     "test": "nrfconnect-scripts test",
     "test-watch": "nrfconnect-scripts test --watch",
     "clean": "npm run clean-dist && npm run clean-modules",
