+++ src/SidePanel/index.scss
